requirements.txt
functions-framework==3.*
google-cloud-secret-manager==2.*
google-cloud-firestore==2.*
google-cloud-logging==3.*
requests==2.*
pydantic==2.*
python-dateutil==2.*
google-generativeai==0.8.*
